Add async refactoring

+++ api/menu_handlers.py
@@ -4,17 +4,16 @@ from fastapi import (
     HTTPException,
     status)
 from pydantic import ValidationError
-from sqlalchemy.orm import Session
-from sqlalchemy.sql import func
+from sqlalchemy.ext.asyncio import AsyncSession
+from sqlalchemy.sql import func, select
 from uuid import uuid4
 
 from db.models import Menu, SubMenu, Dish
-from db.session import get_db
+from db.session import get_session
 from .schemas import (
     MenuCreateRequest,
     MenuCreateResponse,
     MenuPatchRequest)
-from .helpers import check_menu
 
 menu_router = APIRouter()
 
@@ -26,7 +25,7 @@ menu_router = APIRouter()
     status_code=201)
 async def create_menu(
         request: MenuCreateRequest,
-        session: Session = Depends(get_db)):
+        session: AsyncSession = Depends(get_session)):
     try:
         menu = Menu(
             id=str(uuid4()),
@@ -38,15 +37,16 @@ async def create_menu(
             status_code=status.HTTP_400_BAD_REQUEST,
             detail='the data are not valid'
         )
-    if session.query(Menu).filter_by(title=request.title).first():
+    result = await session.execute(select(Menu).filter_by(title=request.title))
+    if result.scalar_one_or_none():
         raise HTTPException(
             status_code=status.HTTP_409_CONFLICT,
             detail='menu already exists'
         )
     session.add(menu)
-    session.commit()
-    session.refresh(menu)
-    session.close()
+    await session.commit()
+    await session.refresh(menu)
+    await session.close()
     return {
         "id": menu.id,
         "title": menu.title,
@@ -55,56 +55,64 @@ async def create_menu(
 
 
 @menu_router.get('/api/v1/menus', tags=['Menus'])
-async def get_menus(session: Session = Depends(get_db)):
-    sub_query = session.query(
+async def get_menus(session: AsyncSession = Depends(get_session)):
+    # TODO: 1 check for errors
+    sub_query = select(
         SubMenu.menu_id,
         func.count(Dish.id).label('total_dishes')
     ).join(
         Dish, Dish.submenu_id == SubMenu.id
     ).group_by(SubMenu.menu_id).subquery()
-    query = session.query(
-        Menu,
-        func.coalesce(func.count(SubMenu.id), 0),
-        func.coalesce(sub_query.c.total_dishes, 0)
-    ).outerjoin(SubMenu, SubMenu.menu_id == Menu.id).outerjoin(
-        sub_query, Menu.id == sub_query.c.menu_id
-    ).group_by(Menu.id).all()
+    query = await session.execute(
+        select(
+            Menu,
+            func.coalesce(func.count(SubMenu.id), 0),
+            func.coalesce(sub_query.c.total_dishes, 0)
+        )
+        .outerjoin(SubMenu, SubMenu.menu_id == Menu.id)
+        .outerjoin(sub_query, Menu.id == sub_query.c.menu_id)
+        .group_by(Menu.id, sub_query.c.total_dishes))
+    result = query.all()
     return [{
         "id": q[0].id,
         "title": q[0].title,
         "description": q[0].description,
         "submenus_count": q[1],
         "dishes_count": q[2]
-    } for q in query]
+    } for q in result]
 
 
 @menu_router.get('/api/v1/menus/{menu_id}', tags=['Menus'])
-async def get_menu_by_id(menu_id: str, session: Session = Depends(get_db)):
-    sub_query = session.query(
+async def get_menu_by_id(menu_id: str, session: AsyncSession = Depends(get_session)):
+    # TODO: 2 check for errors
+    sub_query = select(
         SubMenu.menu_id,
         func.count(Dish.id).label('total_dishes')
     ).join(Dish, Dish.submenu_id == SubMenu.id).group_by(
         SubMenu.menu_id).subquery()
-    query = session.query(
-        Menu,
-        func.coalesce(func.count(SubMenu.id), 0),
-        func.coalesce(sub_query.c.total_dishes, 0)
-    ).filter_by(id=menu_id).outerjoin(
-        SubMenu, SubMenu.menu_id == Menu.id
-    ).outerjoin(
-        sub_query, Menu.id == sub_query.c.menu_id
-    ).group_by(Menu.id).first()
-    if not query:
+    query = await session.execute(
+        select(
+            Menu,
+            func.coalesce(func.count(SubMenu.id), 0),
+            func.coalesce(sub_query.c.total_dishes, 0)
+        ).filter_by(id=menu_id).outerjoin(
+            SubMenu, SubMenu.menu_id == Menu.id
+        ).outerjoin(
+            sub_query, Menu.id == sub_query.c.menu_id
+        ).group_by(Menu.id, sub_query.c.total_dishes)
+    )
+    result = query.first()
+    if not result:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="menu not found"
         )
     return {
-        "id": query[0].id,
-        "title": query[0].title,
-        "description": query[0].description,
-        "submenus_count": query[1],
-        "dishes_count": query[2]
+        "id": result[0].id,
+        "title": result[0].title,
+        "description": result[0].description,
+        "submenus_count": result[1],
+        "dishes_count": result[2]
     }
 
 
@@ -112,20 +120,32 @@ async def get_menu_by_id(menu_id: str, session: Session = Depends(get_db)):
 async def update_menu(
         menu_id: str,
         request: MenuPatchRequest,
-        session: Session = Depends(get_db)):
-    menu = check_menu(session, menu_id)
+        session: AsyncSession = Depends(get_session)):
+    result = await session.execute(select(Menu).filter_by(id=menu_id))
+    menu = result.scalar_one_or_none()
+    if not menu:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="menu not found"
+        )
     if request.title:
         menu.title = request.title
     if request.description:
         menu.description = request.description
-    session.commit()
-    session.refresh(menu)
+    await session.commit()
+    await session.refresh(menu)
     return menu
 
 
 @menu_router.delete('/api/v1/menus/{menu_id}', tags=['Menus'])
-async def delete_menu(menu_id: str, session: Session = Depends(get_db)):
-    menu = check_menu(session, menu_id)
-    session.delete(menu)
-    session.commit()
+async def delete_menu(menu_id: str, session: AsyncSession = Depends(get_session)):
+    result = await session.execute(select(Menu).filter_by(id=menu_id))
+    menu = result.scalar_one_or_none()
+    if not menu:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="menu not found"
+        )
+    await session.delete(menu)
+    await session.commit()
     return {"message": "menu deleted successfully"}
